sicp; revamp the Rakefile
I introduced an Exercise class that makes working with
exercise way easier. Now there is some separation between
the Rake task definitions and the code that operates on
the exercises. It serves two purposes:

1. It makes the Rake tasks way more clear
2. It reduced a bunch of duplication between the tasks

Note: There is a weird return in Exercise.next. I wish FileUtils.cd
returned the result of its block